# routes/exams.js
const express = require('express');
const { authenticateToken } = require('../middleware/auth');
const { authorize } = require('../middleware/rbac');
const asyncHandler = require('../utils/asyncHandler');

const router = express.Router();

// Placeholder: TODO - Implement exams controller
const examsController = {
    listExams: (req, res) => res.json({ success: true, data: [] }),
    createExam: (req, res) => res.status(201).json({ success: true, data: {} }),
    getExam: (req, res) => res.json({ success: true, data: {} }),
    updateExam: (req, res) => res.json({ success: true, data: {} }),
    deleteExam: (req, res) => res.status(204).end()
};

/**
 * @openapi
 * /api/exams:
 *   get:
 *     tags:
 *       - Exams
 *     summary: List exams
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: query
 *         name: page
 *         schema:
 *           type: integer
 *       - in: query
 *         name: limit
 *         schema:
 *           type: integer
 *     responses:
 *       200:
 *         description: OK
 *       401:
 *         description: Unauthorized
 *       403:
 *         description: Forbidden
 */

/**
 * @openapi
 * /api/exams:
 *   post:
 *     tags:
 *       - Exams
 *     summary: Create exam
 *     security:
 *       - bearerAuth: []
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               subject:
 *                 type: string
 *               date:
 *                 type: string
 *                 format: date
 *               duration:
 *                 type: integer
 *     responses:
 *       201:
 *         description: Created
 *       400:
 *         description: Bad Request
 *       401:
 *         description: Unauthorized
 *       403:
 *         description: Forbidden
 */

/**
 * @openapi
 * /api/exams/{id}:
 *   get:
 *     tags:
 *       - Exams
 *     summary: Get exam by id
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: OK
 *       401:
 *         description: Unauthorized
 *       403:
 *         description: Forbidden
 *       404:
 *         description: Not Found
 *   put:
 *     tags:
 *       - Exams
 *     summary: Update exam
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     requestBody:
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               subject:
 *                 type: string
 *               date:
 *                 type: string
 *                 format: date
 *               duration:
 *                 type: integer
 *     responses:
 *       200:
 *         description: Updated
 *       400:
 *         description: Bad Request
 *       401:
 *         description: Unauthorized
 *       403:
 *         description: Forbidden
 *       404:
 *         description: Not Found
 *   delete:
 *     tags:
 *       - Exams
 *     summary: Delete exam
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       204:
 *         description: Deleted
 *       401:
 *         description: Unauthorized
 *       403:
 *         description: Forbidden
 *       404:
 *         description: Not Found
 */

// GET /api/exams - List exams
router.get('/', authenticateToken, authorize('exams', 'read'), asyncHandler(examsController.listExams));

// POST /api/exams - Create exam
router.post('/', authenticateToken, authorize('exams', 'create'), asyncHandler(examsController.createExam));

// GET /api/exams/:id - Get exam
router.get('/:id', authenticateToken, authorize('exams', 'read'), asyncHandler(examsController.getExam));

// PUT /api/exams/:id - Update exam
router.put('/:id', authenticateToken, authorize('exams', 'update'), asyncHandler(examsController.updateExam));

// DELETE /api/exams/:id - Delete exam
router.delete('/:id', authenticateToken, authorize('exams', 'delete'), asyncHandler(examsController.deleteExam));

// POST /api/exams - Create exam
router.post('/', authenticateToken, authorize('exams', 'create'), asyncHandler(examsController.createExam));

// GET /api/exams/:id - Get exam
router.get('/:id', authenticateToken, authorize('exams', 'read'), asyncHandler(examsController.getExam));

// PUT /api/exams/:id - Update exam
router.put('/:id', authenticateToken, authorize('exams', 'update'), asyncHandler(examsController.updateExam));

// DELETE /api/exams/:id - Delete exam
router.delete('/:id', authenticateToken, authorize('exams', 'delete'), asyncHandler(examsController.deleteExam));

module.exports = router;