# routes/fees.js
const express = require('express');
const { authenticateToken } = require('../middleware/auth');
const { authorize } = require('../middleware/rbac');
const asyncHandler = require('../utils/asyncHandler');

const router = express.Router();

// Placeholder: TODO - Implement fees controller
const feesController = {
    listFees: (req, res) => res.json({ success: true, data: [] }),
    createFee: (req, res) => res.status(201).json({ success: true, data: {} }),
    getFee: (req, res) => res.json({ success: true, data: {} }),
    updateFee: (req, res) => res.json({ success: true, data: {} }),
    deleteFee: (req, res) => res.status(204).end()
};

/**
 * @openapi
 * /api/fees:
 *   get:
 *     tags:
 *       - Fees
 *     summary: List fees
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: query
 *         name: page
 *         schema:
 *           type: integer
 *       - in: query
 *         name: limit
 *         schema:
 *           type: integer
 *     responses:
 *       200:
 *         description: OK
 *       401:
 *         description: Unauthorized
 *       403:
 *         description: Forbidden
 */

/**
 * @openapi
 * /api/fees:
 *   post:
 *     tags:
 *       - Fees
 *     summary: Create fee record
 *     security:
 *       - bearerAuth: []
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               studentId:
 *                 type: string
 *               amount:
 *                 type: number
 *               description:
 *                 type: string
 *               dueDate:
 *                 type: string
 *                 format: date
 *     responses:
 *       201:
 *         description: Created
 *       400:
 *         description: Bad Request
 *       401:
 *         description: Unauthorized
 *       403:
 *         description: Forbidden
 */

/**
 * @openapi
 * /api/fees/{id}:
 *   get:
 *     tags:
 *       - Fees
 *     summary: Get fee record by id
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: OK
 *       401:
 *         description: Unauthorized
 *       403:
 *         description: Forbidden
 *       404:
 *         description: Not Found
 *   put:
 *     tags:
 *       - Fees
 *     summary: Update fee record
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     requestBody:
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               amount:
 *                 type: number
 *               description:
 *                 type: string
 *               dueDate:
 *                 type: string
 *                 format: date
 *     responses:
 *       200:
 *         description: Updated
 *       400:
 *         description: Bad Request
 *       401:
 *         description: Unauthorized
 *       403:
 *         description: Forbidden
 *       404:
 *         description: Not Found
 *   delete:
 *     tags:
 *       - Fees
 *     summary: Delete fee record
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       204:
 *         description: Deleted
 *       401:
 *         description: Unauthorized
 *       403:
 *         description: Forbidden
 *       404:
 *         description: Not Found
 */

// GET /api/fees - List fees
router.get('/', authenticateToken, authorize('fees', 'read'), asyncHandler(feesController.listFees));

// POST /api/fees - Create fee record
router.post('/', authenticateToken, authorize('fees', 'create'), asyncHandler(feesController.createFee));

// GET /api/fees/:id - Get fee record
router.get('/:id', authenticateToken, authorize('fees', 'read'), asyncHandler(feesController.getFee));

// PUT /api/fees/:id - Update fee record
router.put('/:id', authenticateToken, authorize('fees', 'update'), asyncHandler(feesController.updateFee));

// DELETE /api/fees/:id - Delete fee record
router.delete('/:id', authenticateToken, authorize('fees', 'delete'), asyncHandler(feesController.deleteFee));

// POST /api/fees - Create fee record
router.post('/', authenticateToken, authorize('fees', 'create'), asyncHandler(feesController.createFee));

// GET /api/fees/:id - Get fee record
router.get('/:id', authenticateToken, authorize('fees', 'read'), asyncHandler(feesController.getFee));

// PUT /api/fees/:id - Update fee record
router.put('/:id', authenticateToken, authorize('fees', 'update'), asyncHandler(feesController.updateFee));

// DELETE /api/fees/:id - Delete fee record
router.delete('/:id', authenticateToken, authorize('fees', 'delete'), asyncHandler(feesController.deleteFee));

module.exports = router;